Restore mark back to chart and tool tip
