+++ src/components/Editors/Chart/Sections/Chart.tsx
@@ -5,7 +5,8 @@ import SelectField from '../../../Parts/Fields/Select'
 import EditorSection from '../../../Parts/Editor/Section'
 import Columns from '../../../Parts/Columns'
 import { useStore, selectors } from '../store'
-import Mark from './Mark'
+import * as settings from '../settings'
+import YesNoField from '../../../Parts/Fields/YesNo'
 
 export default function Chart() {
   const updateHelp = useStore((state) => state.updateHelp)
@@ -14,13 +15,14 @@ export default function Chart() {
       <Columns spacing={3}>
         <Box>
           <Table />
-          <Height />
+          <Mark />
+          <Tooltip />
         </Box>
         <Box>
           <Width />
+          <Height />
         </Box>
       </Columns>
-      <Mark />
     </EditorSection>
   )
 }
@@ -73,6 +75,55 @@ function Height() {
       value={height || ''}
       onFocus={() => updateHelp('chart/height')}
       onChange={(value) => updateDescriptor({ height: parseInt(value) || undefined })}
+    />
+  )
+}
+
+function Mark() {
+  const descriptor = useStore((state) => state.descriptor)
+  const mark = useStore((state) => state.descriptor.mark)
+  const updateHelp = useStore((state) => state.updateHelp)
+  const updateState = useStore((state) => state.updateState)
+  const type = typeof mark === 'object' ? mark.type : mark
+  return (
+    <SelectField
+      label="Mark"
+      value={type || ''}
+      options={settings.MARKS}
+      onFocus={() => updateHelp('chart/mark')}
+      onChange={(value) => {
+        if (!value) return
+        descriptor.mark = value
+        updateState({ descriptor })
+      }}
+    />
+  )
+}
+
+function Tooltip() {
+  const descriptor = useStore((state) => state.descriptor)
+  const mark = useStore((state) => state.descriptor.mark)
+  const updateHelp = useStore((state) => state.updateHelp)
+  const updateState = useStore((state) => state.updateState)
+  const tooltip = typeof mark === 'object' ? mark.tooltip : settings.DEFAULT_TOOLTIP
+  return (
+    <YesNoField
+      label="Tooltip"
+      value={tooltip}
+      onFocus={() => updateHelp('chart/markTooltip')}
+      onChange={(value) => {
+        const type = typeof mark === 'string' ? mark : mark?.type
+        if (value) {
+          descriptor.mark = { type: type ?? '', tooltip: value }
+        } else {
+          if (type) {
+            descriptor.mark = type
+          } else {
+            delete descriptor.mark
+          }
+        }
+        updateState({ descriptor })
+      }}
     />
   )
 }
